Obey clippy

+++ tremor-script/src/std_lib/array.rs
@@ -30,7 +30,7 @@ pub fn load(registry: &mut Registry) {
             Ok(Value::from(output))
         }))
         .insert(tremor_const_fn! (array|len(_context, _input: Array) {
-            Ok(Value::from(_input.len() as i64))
+            Ok(Value::from(_input.len()))
         }))
         .insert(tremor_const_fn! (array|is_empty(_context, _input: Array) {
             Ok(Value::from(_input.is_empty()))
